move some plugins to components
