+++ artview/components/gatefilter.py
