+++ artview/components/manual_edit.py
