+++ artview/modes.py
@@ -354,6 +354,23 @@ def radar_terminal_mode():
         ]
     )
 
+
+def cmap_radar_mode():
+    change_mode(
+        [FileNavigator, RadarDisplay, ColormapEdit],
+        [
+            ((1, 'Vcolormap'), (2, 'Vcolormap')),
+            ]
+        )
+
+def cmap_grid_mode():
+    change_mode(
+        [FileNavigator, GridDisplay, ColormapEdit],
+        [
+            ((1, 'Vcolormap'), (2, 'Vcolormap')),
+            ]
+        )
+
 modes = [
     {'label': 'Add RadarDisplay',
      'group': 'graph',
@@ -400,4 +417,10 @@ modes = [
     {'label': 'Manipulate Radar in Terminal',
      'group': 'terminal',
      'action': radar_terminal_mode},
+    {'label': 'Edit Colormap (radar)',
+     'group': 'graph',
+     'action': cmap_radar_mode},
+    {'label': 'Edit Colormap (grid)',
+     'group': 'graph',
+     'action': cmap_grid_mode},
     ]
