--- artview/plugins/mapper.py
@@ -1,701 +1,0 @@
-"""
-mapper.py
-"""
-from __future__ import print_function
-# Load the needed packages
-from functools import partial
-
-import pyart
-from pyart.config import get_field_name
-import time
-import os
-
-from ..core import (Component, Variable, common, QtWidgets, QtGui,
-                    QtCore, VariableChoose, log)
-
-class Mapper(Component):
-    '''
-    Interface for executing :py:func:`pyart.map.grid_from_radars`
-    '''
-
-    Vradar = None  #: see :ref:`shared_variable`
-    Vgrid = None  #: see :ref:`shared_variable`
-
-    @classmethod
-    def guiStart(self, parent=None):
-        '''Graphical interface for starting this class'''
-        kwargs, independent = \
-            common._SimplePluginStart("Mapper").startDisplay()
-        kwargs['parent'] = parent
-        return self(**kwargs), independent
-
-    def __init__(self, Vradar=None, Vgrid=None, name="Mapper", parent=None):
-        '''Initialize the class to create the interface.
-
-        Parameters
-        ----------
-        [Optional]
-        Vradar : :py:class:`~artview.core.core.Variable` instance
-            Radar signal variable.
-            A value of None initializes an empty Variable.
-        Vgrid : :py:class:`~artview.core.core.Variable` instance
-            Grid signal variable.
-            A value of None initializes an empty Variable.
-        name : string
-            Field Radiobutton window name.
-        parent : PyQt instance
-            Parent instance to associate to this class.
-            If None, then Qt owns, otherwise associated w/ parent PyQt instance
-        '''
-        super(Mapper, self).__init__(name=name, parent=parent)
-        self.central_widget = QtWidgets.QWidget()
-        self.setCentralWidget(self.central_widget)
-        self.layout = QtWidgets.QGridLayout(self.central_widget)
-
-        self.mountUI()
-
-        self.parameters = {
-            "radars": None,
-            "gridshape": (1, 500, 500),
-            "grid_limits": (
-                (2000, 3000),
-                (-250000, 250000),
-                (-250000, 250000)),
-            "grid_origin": (0, 0),
-            "grid_origin_lat": 0,
-            "grid_origin_lon": 0,
-            "grid_origin_alt": 0,
-            "gridding_algo": 'map_gates_to_grid',
-            "fields": [],
-            "refl_filter_flag": True,
-            "refl_field": get_field_name('reflectivity'),
-            "max_refl": 100,
-            "map_roi": True,
-            "weighting_function": "Barnes",
-            "toa": 17000,
-            "roi_func": "dist_beam",
-            "constant_roi": 500,
-            "z_factor": 0.05,
-            "xy_factor": 0.02,
-            "min_radius": 500,
-            "bsp": 1,
-            "copy_field_data": True,
-            "algorithm": "kd_tree",
-            "leafsize": 10,
-            }
-
-        self.general_parameters_type = [
-            ("grid_origin_lat", float),
-            ("grid_origin_lon", float),
-            ("grid_origin_alt", float),
-            ("gridding_algo", ('map_to_grid', 'map_gates_to_grid')),
-            ("refl_filter_flag", bool),
-            ("refl_field", str),
-            ("max_refl", float),
-            ("map_roi", bool),
-            ("weighting_function", ("Barnes", "Cressman")),
-            ("toa", float),
-            ]
-        self.roi_parameters_type = [
-            ("roi_func", ("constant", "dist", "dist_beam")),
-            ("constant_roi", float),
-            ("z_factor", float),
-            ("xy_factor", float),
-            ("min_radius", float),
-            ("bsp", float)
-            ]
-        self.gridding_parameters_type = [
-            ("copy_field_data", bool),
-            ("algorithm", ("kd_tree", "ball_tree")),
-            ("leafsize", int),
-            ]
-
-        if Vradar is None:
-            self.Vradar = Variable(None)
-        else:
-            self.Vradar = Vradar
-
-        if Vgrid is None:
-            self.Vgrid = Variable(None)
-        else:
-            self.Vgrid = Vgrid
-        self.sharedVariables = {"Vradar": self.NewRadar,
-                                "Vgrid": None}
-        self.connectAllVariables()
-
-        self.NewRadar(None, True)
-        self.show()
-
-    def mountUI(self):
-        '''Mount Options Layout.'''
-
-        self.despeckleButton = QtWidgets.QPushButton("Map")
-        self.despeckleButton.clicked.connect(self.grid_from_radars)
-        self.layout.addWidget(self.despeckleButton, 7, 0, 1, 2)
-
-        parentdir = os.path.abspath(os.path.join(os.path.dirname(__file__),
-                                                 os.pardir))
-        config_icon = QtGui.QIcon(os.sep.join(
-            [parentdir, 'icons', "categories-applications-system-icon.png"]))
-        self.configButton = QtWidgets.QPushButton(config_icon,"")
-        self.layout.addWidget(self.configButton, 7, 2)
-        self.configMenu = QtWidgets.QMenu(self)
-        self.configButton.setMenu(self.configMenu)
-
-        self.configMenu.addAction(QtWidgets.QAction("Set General Parameters", self,
-                                                triggered=self.setParameters))
-        self.configMenu.addAction(QtWidgets.QAction("Set Radius of Influence Parameters", self,
-                                                triggered=self.setRoiParameters))
-        self.configMenu.addAction(QtWidgets.QAction("Set map_to_grid Parameters", self,
-                                                triggered=self.setGriddingParameters))
-        self.fieldMenu = self.configMenu.addMenu("Fields")
-        self.configMenu.addAction(QtWidgets.QAction("Help", self,
-                                                triggered=self._displayHelp))
-
-        self.layout.addWidget(QtWidgets.QLabel("Z"), 1, 0, 2, 1)
-        self.layout.addWidget(QtWidgets.QLabel("Y"), 3, 0, 2, 1)
-        self.layout.addWidget(QtWidgets.QLabel("X"), 5, 0, 2, 1)
-
-        self.gridShapeZ = QtWidgets.QSpinBox()
-        self.gridShapeZ.setRange(0, 1000000)
-        self.gridShapeZ.setValue(1)
-        self.gridShapeY = QtWidgets.QSpinBox()
-        self.gridShapeY.setRange(0, 1000000)
-        self.gridShapeY.setValue(500)
-        self.gridShapeX = QtWidgets.QSpinBox()
-        self.gridShapeX.setRange(0, 1000000)
-        self.gridShapeX.setValue(500)
-        self.layout.addWidget(QtWidgets.QLabel("grid_shape"), 0, 1)
-        self.layout.addWidget(self.gridShapeZ, 1, 1, 2, 1)
-        self.layout.addWidget(self.gridShapeY, 3, 1, 2, 1)
-        self.layout.addWidget(self.gridShapeX, 5, 1, 2, 1)
-
-        self.gridLimitsZmin = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsZmin.setRange(-41000000, 41000000)
-        self.gridLimitsZmin.setSingleStep(1000)
-        self.gridLimitsZmin.setValue(2000)
-        self.gridLimitsZmax = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsZmax.setRange(-41000000, 41000000)
-        self.gridLimitsZmax.setSingleStep(1000)
-        self.gridLimitsZmax.setValue(3000)
-        self.gridLimitsYmin = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsYmin.setRange(-41000000, 41000000)
-        self.gridLimitsYmin.setSingleStep(1000)
-        self.gridLimitsYmin.setValue(-250000)
-        self.gridLimitsYmax = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsYmax.setRange(-41000000, 41000000)
-        self.gridLimitsYmax.setSingleStep(1000)
-        self.gridLimitsYmax.setValue(250000)
-        self.gridLimitsXmin = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsXmin.setRange(-41000000, 41000000)
-        self.gridLimitsXmin.setSingleStep(1000)
-        self.gridLimitsXmin.setValue(-250000)
-        self.gridLimitsXmax = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsXmax.setRange(-41000000, 41000000)
-        self.gridLimitsXmax.setSingleStep(1000)
-        self.gridLimitsXmax.setValue(250000)
-        self.layout.addWidget(QtWidgets.QLabel("grid_limits (m)"), 0, 2)
-        self.layout.addWidget(self.gridLimitsZmin, 1, 2)
-        self.layout.addWidget(self.gridLimitsZmax, 2, 2)
-        self.layout.addWidget(self.gridLimitsYmin, 3, 2)
-        self.layout.addWidget(self.gridLimitsYmax, 4, 2)
-        self.layout.addWidget(self.gridLimitsXmin, 5, 2)
-        self.layout.addWidget(self.gridLimitsXmax, 6, 2)
-
-        self.layout.setRowStretch(8, 1)
-        self.layout.setColumnStretch(1, 1)
-        self.layout.setColumnStretch(2, 1)
-
-    def grid_from_radars(self):
-        '''Mount Options and execute :py:func:`~pyart.correct.grid_from_radars`.
-        The resulting grid is update in Vgrid.
-        '''
-        # test radar
-        if self.Vradar.value is None:
-            common.ShowWarning("Radar is None, can not perform correction")
-            return
-        # mount options
-        self.parameters['radars'] = (self.Vradar.value,)
-        self.parameters['fields'] = []
-        for field in self.field_actions.keys():
-            if self.field_actions[field].isChecked():
-                self.parameters['fields'].append(field)
-
-        self.parameters['grid_shape'] = (self.gridShapeZ.value(),
-                                         self.gridShapeY.value(),
-                                         self.gridShapeX.value())
-        self.parameters['grid_limits'] = (
-            (self.gridLimitsZmin.value(), self.gridLimitsZmax.value()),
-            (self.gridLimitsYmin.value(), self.gridLimitsYmax.value()),
-            (self.gridLimitsXmin.value(), self.gridLimitsXmax.value()))
-
-        self.parameters['grid_origin'] = (self.parameters['grid_origin_lat'],
-                                          self.parameters['grid_origin_lon'])
-
-        # execute
-        print("mapping ..", file=log.debug)
-        t0 = time.time()
-        try:
-            grid = pyart.map.grid_from_radars(**self.parameters)
-        except:
-            import traceback
-            error = traceback.format_exc()
-            common.ShowLongText("Py-ART fails with following error\n\n" +
-                                error)
-        t1 = time.time()
-        print("Mapping took %fs" % (t1-t0), file=log.debug)
-
-        # update
-        self.Vgrid.change(grid)
-
-    def setParameters(self):
-        '''Open set parameters dialog.'''
-        parm = common.get_options(self.general_parameters_type, self.parameters)
-        for key in parm.keys():
-            self.parameters[key] = parm[key]
-
-    def setGriddingParameters(self):
-        '''Open set parameters dialog.'''
-        parm = common.get_options(self.gridding_parameters_type, self.parameters)
-        for key in parm.keys():
-            self.parameters[key] = parm[key]
-
-    def setRoiParameters(self):
-        '''Open set parameters dialog.'''
-        parm = common.get_options(self.roi_parameters_type, self.parameters)
-        for key in parm.keys():
-            self.parameters[key] = parm[key]
-
-    def _displayHelp(self):
-        '''Display Py-Art's docstring for help.'''
-        common.ShowLongText(pyart.map.grid_from_radars.__doc__)
-
-    def NewRadar(self, variable, strong):
-        '''Display Py-Art's docstring for help.'''
-        if self.Vradar.value is None:
-            return
-
-        fields = self.Vradar.value.fields.keys()
-        self.field_radio_group = QtWidgets.QActionGroup(self, exclusive=False)
-        self.field_actions = {}
-        self.fieldMenu.clear()
-        for field in fields:
-            action = self.field_radio_group.addAction(field)
-            action.setCheckable(True)
-            action.setChecked(True)
-            self.fieldMenu.addAction(action)
-            self.field_actions[field] = action
-
-        lat = float(self.Vradar.value.latitude['data'])
-        lon = float(self.Vradar.value.longitude['data'])
-        alt = float(self.Vradar.value.altitude['data'])
-        self.parameters["grid_origin_lat"] = lat
-        self.parameters["grid_origin_lon"] = lon
-        self.parameters["grid_origin_alt"] = alt
-
-class Mapper_old(Component):
-    '''
-    Interface for executing :py:func:`pyart.map.grid_from_radars`
-    '''
-
-    Vradar = None  #: see :ref:`shared_variable`
-    Vgrid = None  #: see :ref:`shared_variable`
-
-    @classmethod
-    def guiStart(self, parent=None):
-        '''Graphical interface for starting this class'''
-        kwargs, independent = \
-            common._SimplePluginStart("Mapper").startDisplay()
-        kwargs['parent'] = parent
-        return self(**kwargs), independent
-
-    def __init__(self, Vradar=None, Vgrid=None, name="Mapper", parent=None):
-        '''Initialize the class to create the interface.
-
-        Parameters
-        ----------
-        [Optional]
-        Vradar : :py:class:`~artview.core.core.Variable` instance
-            Radar signal variable.
-            A value of None initializes an empty Variable.
-        Vgrid : :py:class:`~artview.core.core.Variable` instance
-            Grid signal variable.
-            A value of None initializes an empty Variable.
-        name : string
-            Field Radiobutton window name.
-        parent : PyQt instance
-            Parent instance to associate to this class.
-            If None, then Qt owns, otherwise associated w/ parent PyQt instance
-        '''
-        super(Mapper_old, self).__init__(name=name, parent=parent)
-        self.central_widget = QtWidgets.QWidget()
-        self.setCentralWidget(self.central_widget)
-
-        if Vradar is None:
-            self.Vradar = Variable(None)
-        else:
-            self.Vradar = Vradar
-
-        if Vgrid is None:
-            self.Vgrid = Variable(None)
-        else:
-            self.Vgrid = Vgrid
-        self.sharedVariables = {"Vradar": self.NewRadar,
-                                "Vgrid": None}
-        self.connectAllVariables()
-
-        self.generalLayout = QtWidgets.QGridLayout()
-        self.especificLayout = QtWidgets.QGridLayout()
-        self.layout.addLayout(self.generalLayout)
-        self.layout.addLayout(self.especificLayout)
-
-        self.button = QtWidgets.QPushButton("Map")
-        self.button.clicked.connect(self.grid_from_radars)
-        self.button.setToolTip('Execute pyart.map.grid_from_radars')
-        self.layout.addWidget(self.button)
-
-        self.addGeneralOptions()
-
-        self.NewRadar(None, True)
-
-        self.show()
-
-    def addGeneralOptions(self):
-        '''Mount Options Layout for :py:func:`~pyart.map.grid_from_radars`'''
-        self.generalLayout.addWidget(QtWidgets.QLabel("Z"), 0, 1, 1, 2)
-        self.generalLayout.addWidget(QtWidgets.QLabel("Y"), 0, 3, 1, 2)
-        self.generalLayout.addWidget(QtWidgets.QLabel("X"), 0, 5, 1, 2)
-
-        self.gridShapeZ = QtWidgets.QSpinBox()
-        self.gridShapeZ.setRange(0, 1000000)
-        self.gridShapeZ.setValue(1)
-        self.gridShapeY = QtWidgets.QSpinBox()
-        self.gridShapeY.setRange(0, 1000000)
-        self.gridShapeY.setValue(500)
-        self.gridShapeX = QtWidgets.QSpinBox()
-        self.gridShapeX.setRange(0, 1000000)
-        self.gridShapeX.setValue(500)
-        self.generalLayout.addWidget(QtWidgets.QLabel("grid_shape"), 1, 0)
-        self.generalLayout.addWidget(self.gridShapeZ, 1, 1, 1, 2)
-        self.generalLayout.addWidget(self.gridShapeY, 1, 3, 1, 2)
-        self.generalLayout.addWidget(self.gridShapeX, 1, 5, 1, 2)
-
-        self.gridLimitsZmin = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsZmin.setRange(-41000000, 41000000)
-        self.gridLimitsZmin.setSingleStep(1000)
-        self.gridLimitsZmin.setValue(2000)
-        self.gridLimitsZmax = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsZmax.setRange(-41000000, 41000000)
-        self.gridLimitsZmax.setSingleStep(1000)
-        self.gridLimitsZmax.setValue(3000)
-        self.gridLimitsYmin = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsYmin.setRange(-41000000, 41000000)
-        self.gridLimitsYmin.setSingleStep(1000)
-        self.gridLimitsYmin.setValue(-250000)
-        self.gridLimitsYmax = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsYmax.setRange(-41000000, 41000000)
-        self.gridLimitsYmax.setSingleStep(1000)
-        self.gridLimitsYmax.setValue(250000)
-        self.gridLimitsXmin = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsXmin.setRange(-41000000, 41000000)
-        self.gridLimitsXmin.setSingleStep(1000)
-        self.gridLimitsXmin.setValue(-250000)
-        self.gridLimitsXmax = QtWidgets.QDoubleSpinBox()
-        self.gridLimitsXmax.setRange(-41000000, 41000000)
-        self.gridLimitsXmax.setSingleStep(1000)
-        self.gridLimitsXmax.setValue(250000)
-        self.generalLayout.addWidget(QtWidgets.QLabel("grid_limits"), 2, 0)
-        self.generalLayout.addWidget(self.gridLimitsZmin, 2, 1)
-        self.generalLayout.addWidget(self.gridLimitsZmax, 2, 2)
-        self.generalLayout.addWidget(self.gridLimitsYmin, 2, 3)
-        self.generalLayout.addWidget(self.gridLimitsYmax, 2, 4)
-        self.generalLayout.addWidget(self.gridLimitsXmin, 2, 5)
-        self.generalLayout.addWidget(self.gridLimitsXmax, 2, 6)
-
-        self.griddingAlgo = QtWidgets.QComboBox()
-        self.griddingAlgo.addItem('map_to_grid')
-        self.griddingAlgo.addItem('map_gates_to_grid')
-        self.generalLayout.addWidget(QtWidgets.QLabel("gridding_algo"), 3, 0)
-        self.generalLayout.addWidget(self.griddingAlgo, 3, 1, 1, 6)
-
-        self.griddingAlgo.currentIndexChanged[str].connect(
-            self.addEspecificOptions)
-        self.griddingAlgo.setCurrentIndex(1)
-
-    def addEspecificOptions(self, item):
-        '''Mount Options Layout depending on gridding algorithm.'''
-        self._clearLayout(self.especificLayout)
-        if item == 'map_gates_to_grid':
-            self.mapGatesToGridOptions()
-        elif item == 'map_to_grid':
-            self.mapToGridOptions()
-        self.NewRadar(None, True)
-
-    def mapGatesToGridOptions(self):
-        '''Mount Options Layout for :py:func:`~pyart.map.map_gates_to_grid`'''
-        self._fieldsOptions()
-        self._interpolationOptions()
-
-    def mapToGridOptions(self):
-        '''Mount Options Layout for :py:func:`~pyart.map.map_to_grid`'''
-        self._fieldsOptions()
-        self._interpolationOptions()
-
-        self.copyFieldData = QtWidgets.QCheckBox("copy_field_data")
-        self.copyFieldData.setChecked(True)
-        self.especificLayout.addWidget(self.copyFieldData, 11, 1, 1, 2)
-
-        self.algorithm = QtWidgets.QComboBox()
-        self.algorithm.addItem('kd_tree')
-        self.algorithm.addItem('ball_tree')
-        self.algorithm.setCurrentIndex(0)
-        self.especificLayout.addWidget(QtWidgets.QLabel("algorithm"), 12, 0)
-        self.especificLayout.addWidget(self.algorithm, 12, 1, 1, 2)
-
-        self.leafsize = QtWidgets.QSpinBox()
-        self.leafsize.setRange(0, 1000000)
-        self.leafsize.setValue(10)
-        self.especificLayout.addWidget(QtWidgets.QLabel("leafsize"), 13, 0)
-        self.especificLayout.addWidget(self.leafsize, 13, 1, 1, 2)
-
-    def _fieldsOptions(self):
-        '''Mount Options Layout related to field.'''
-        self.gridOriginLat = QtWidgets.QDoubleSpinBox()
-        self.gridOriginLat.setRange(-90, 90)
-        self.gridOriginLat.setDecimals(8)
-        self.gridOriginLon = QtWidgets.QDoubleSpinBox()
-        self.gridOriginLon.setRange(-180, 180)
-        self.gridOriginLon.setDecimals(8)
-        self.especificLayout.addWidget(QtWidgets.QLabel("grid_origin"), 0, 0)
-        self.especificLayout.addWidget(self.gridOriginLat, 0, 1)
-        self.especificLayout.addWidget(self.gridOriginLon, 0, 2)
-
-        self.gridOriginAlt = QtWidgets.QDoubleSpinBox()
-        self.gridOriginAlt.setRange(0, 20000)
-        self.especificLayout.addWidget(QtWidgets.QLabel("grid_origin_alt"), 1, 0)
-        self.especificLayout.addWidget(self.gridOriginAlt, 1, 1, 1, 2)
-
-        self.fieldsbutton = QtWidgets.QToolButton(self)
-        self.fieldsbutton.setText('Select Fields')
-        self.fieldsmenu = QtWidgets.QMenu(self)
-        self.fieldsbutton.setMenu(self.fieldsmenu)
-        self.fieldsbutton.setPopupMode(QtWidgets.QToolButton.InstantPopup)
-        self.especificLayout.addWidget(self.fieldsbutton, 2, 1, 1, 2)
-
-        self.reflFilterFlag = QtWidgets.QCheckBox("refl_filter_flag")
-        self.reflFilterFlag.setChecked(True)
-        self.especificLayout.addWidget(self.reflFilterFlag, 3, 1, 1, 2)
-
-        self.reflField = QtWidgets.QLineEdit("reflectivity")
-        self.especificLayout.addWidget(QtWidgets.QLabel("refl_field"), 4, 0)
-        self.especificLayout.addWidget(self.reflField, 4, 1, 1, 2)
-
-        self.maxRefl = QtWidgets.QDoubleSpinBox()
-        self.maxRefl.setRange(-1000, 1000)
-        self.maxRefl.setValue(100)
-        self.especificLayout.addWidget(QtWidgets.QLabel("max_refl"), 5, 0)
-        self.especificLayout.addWidget(self.maxRefl, 5, 1, 1, 2)
-
-    def _interpolationOptions(self):
-        '''Mount Options Layout related to interpolation.'''
-        self.mapRoi = QtWidgets.QCheckBox("map_roi")
-        self.mapRoi.setChecked(True)
-        self.especificLayout.addWidget(self.mapRoi, 6, 1, 1, 2)
-
-        self.weightingFunction = QtWidgets.QComboBox()
-        self.weightingFunction.addItem('Barnes')
-        self.weightingFunction.addItem('Cressman')
-        self.weightingFunction.setCurrentIndex(0)
-        self.especificLayout.addWidget(
-            QtWidgets.QLabel("weighting_function"), 7, 0)
-        self.especificLayout.addWidget(self.weightingFunction, 7, 1, 1, 2)
-
-        self.toa = QtWidgets.QDoubleSpinBox()
-        self.toa.setRange(0, 30000)
-        self.toa.setValue(17000)
-        self.toa.setSingleStep(1000)
-        self.especificLayout.addWidget(QtWidgets.QLabel("toa"), 8, 0)
-        self.especificLayout.addWidget(self.toa, 8, 1, 1, 2)
-
-        self.roiFunc = QtWidgets.QComboBox()
-        self.roiFunc.addItem('constant')
-        self.roiFunc.addItem('dist')
-        self.roiFunc.addItem('dist_beam')
-        self.especificLayout.addWidget(QtWidgets.QLabel("roi_func"), 9, 0)
-        self.especificLayout.addWidget(self.roiFunc, 9, 1, 1, 2)
-
-        self.roiFuncLayout = QtWidgets.QGridLayout()
-        self.especificLayout.addLayout(self.roiFuncLayout, 10, 0, 1, 3)
-
-        self.roiFunc.currentIndexChanged[str].connect(self._roiFuncOptions)
-        self.roiFunc.setCurrentIndex(2)
-
-    def _roiFuncOptions(self, item):
-        '''Mount Options Layout related to radius of influence.'''
-        self._clearLayout(self.roiFuncLayout)
-        if item == 'constant':
-            self.constantRoiOptions()
-        elif item == 'dist':
-            self.distRoiOptions()
-        elif item == 'dist_beam':
-            self.distBeamRoiOptions()
-
-    def constantRoiOptions(self):
-        '''Mount Options Layout for constant radius of influence.'''
-        self.constantRoi = QtWidgets.QDoubleSpinBox()
-        self.constantRoi.setRange(0, 30000)
-        self.constantRoi.setValue(500)
-        self.constantRoi.setSingleStep(100)
-        self.roiFuncLayout.addWidget(QtWidgets.QLabel("constant_roi"), 0, 0)
-        self.roiFuncLayout.addWidget(self.constantRoi, 0, 1, 1, 2)
-
-    def distRoiOptions(self):
-        '''Mount Options Layout for dist radius of influence.'''
-        self.zFactor = QtWidgets.QDoubleSpinBox()
-        self.zFactor.setRange(0, 30000)
-        self.zFactor.setValue(0.05)
-        self.zFactor.setSingleStep(0.01)
-        self.zFactor.setDecimals(3)
-        self.roiFuncLayout.addWidget(QtWidgets.QLabel("z_factor"), 0, 0)
-        self.roiFuncLayout.addWidget(self.zFactor, 0, 1, 1, 2)
-
-        self.xyFactor = QtWidgets.QDoubleSpinBox()
-        self.xyFactor.setRange(0, 30000)
-        self.xyFactor.setValue(0.02)
-        self.xyFactor.setSingleStep(0.01)
-        self.xyFactor.setDecimals(3)
-        self.roiFuncLayout.addWidget(QtWidgets.QLabel("xy_factor"), 1, 0)
-        self.roiFuncLayout.addWidget(self.xyFactor, 1, 1, 1, 2)
-
-        self.minRadius = QtWidgets.QDoubleSpinBox()
-        self.minRadius.setRange(0, 30000)
-        self.minRadius.setValue(500)
-        self.minRadius.setSingleStep(100)
-        self.roiFuncLayout.addWidget(QtWidgets.QLabel("min_radius"), 2, 0)
-        self.roiFuncLayout.addWidget(self.minRadius, 2, 1, 1, 2)
-
-    def distBeamRoiOptions(self):
-        '''Mount Options Layout for dist beam radius of influence.'''
-        self.hFactor = QtWidgets.QDoubleSpinBox()
-        self.hFactor.setRange(0, 30000)
-        self.hFactor.setValue(1.0)
-        self.hFactor.setSingleStep(0.1)
-        self.roiFuncLayout.addWidget(QtWidgets.QLabel("h_factor"), 0, 0)
-        self.roiFuncLayout.addWidget(self.hFactor, 0, 1, 1, 2)
-
-        self.nb = QtWidgets.QDoubleSpinBox()
-        self.nb.setRange(0, 30000)
-        self.nb.setValue(1.5)
-        self.nb.setSingleStep(0.1)
-        self.roiFuncLayout.addWidget(QtWidgets.QLabel("nb"), 1, 0)
-        self.roiFuncLayout.addWidget(self.nb, 1, 1, 1, 2)
-
-        self.bsp = QtWidgets.QDoubleSpinBox()
-        self.bsp.setRange(0, 30000)
-        self.bsp.setValue(1.0)
-        self.bsp.setSingleStep(0.1)
-        self.roiFuncLayout.addWidget(QtWidgets.QLabel("bsp"), 2, 0)
-        self.roiFuncLayout.addWidget(self.bsp, 2, 1, 1, 2)
-
-        self.minRadius = QtWidgets.QDoubleSpinBox()
-        self.minRadius.setRange(0, 30000)
-        self.minRadius.setValue(500)
-        self.minRadius.setSingleStep(100)
-        self.roiFuncLayout.addWidget(QtWidgets.QLabel("min_radius"), 3, 0)
-        self.roiFuncLayout.addWidget(self.minRadius, 3, 1, 1, 2)
-
-    def NewRadar(self, variable, strong):
-        '''Display Py-Art's docstring for help.'''
-        if self.Vradar.value is None:
-            return
-        self.fieldsmenu.clear()
-        for field in self.Vradar.value.fields.keys():
-            action = self.fieldsmenu.addAction(field)
-            action.setCheckable(True)
-            action.setChecked(True)
-        lat = float(self.Vradar.value.latitude['data'])
-        lon = float(self.Vradar.value.longitude['data'])
-        alt = float(self.Vradar.value.altitude['data'])
-        self.gridOriginLat.setValue(lat)
-        self.gridOriginLon.setValue(lon)
-        self.gridOriginAlt.setValue(alt)
-
-    def grid_from_radars(self):
-        '''Mount Options and execute :py:func:`~pyart.correct.grid_from_radars`.
-        The resulting grid is update in Vgrid.
-        '''
-        # test radar
-        if self.Vradar.value is None:
-            common.ShowWarning("Radar is None, can not perform correction")
-            return
-        # mount options
-        args = {
-            'radars': (self.Vradar.value,),
-            'grid_shape': (self.gridShapeZ.value(),
-                           self.gridShapeY.value(),
-                           self.gridShapeX.value()),
-            'grid_limits': (
-                (self.gridLimitsZmin.value(), self.gridLimitsZmax.value()),
-                (self.gridLimitsYmin.value(), self.gridLimitsYmax.value()),
-                (self.gridLimitsXmin.value(), self.gridLimitsXmax.value())),
-            'grid_origin': (self.gridOriginLat.value(),
-                            self.gridOriginLon.value()),
-            'gridding_algo': str(self.griddingAlgo.currentText()),
-            'grid_origin_alt': self.gridOriginAlt.value(),
-            'fields': [str(a.text()) for a in
-                       self.fieldsmenu.actions() if a.isChecked()],
-            'refl_filter_flag': self.reflFilterFlag.isChecked(),
-            'refl_field': str(self.reflField.text()),
-            'max_refl': self.maxRefl.value(),
-            'map_roi': self.mapRoi.isChecked(),
-            'weighting_function': str(self.weightingFunction.currentText()),
-            'toa': self.toa.value(),
-            'roi_func': str(self.roiFunc.currentText()),
-        }
-
-        if args['roi_func'] == 'constant':
-            args['constant_roi'] = self.constantRoi.value()
-        elif args['roi_func'] == 'dist':
-            args['z_factor'] = self.zFactor.value()
-            args['xy_factor'] = self.xyFactor.value()
-            args['min_radius'] = self.minRadius.value()
-        elif args['roi_func'] == 'dist_beam':
-            args['h_factor'] = self.hFactor.value()
-            args['nb'] = self.nb.value()
-            args['bsp'] = self.bsp.value()
-            args['min_radius'] = self.minRadius.value()
-
-        if args['gridding_algo'] == 'map_to_grid':
-            args['copy_field_data'] = self.copyFieldData.isChecked()
-            args['algorithm'] = str(self.algorithm.currentText())
-            args['leafsize'] = self.leafsize.value()
-
-        print(args)
-
-        # execute
-        print("mapping ..")
-        t0 = time.time()
-        try:
-            grid = pyart.map.grid_from_radars(**args)
-        except:
-            import traceback
-            error = traceback.format_exc()
-            common.ShowLongText("Py-ART fails with following error\n\n" +
-                                error)
-        t1 = time.time()
-        print(("Mapping took %fs" % (t1-t0)))
-
-        # update
-        self.Vgrid.change(grid)
-
-    def _clearLayout(self, layout):
-        '''recursively remove items from layout.'''
-        while layout.count():
-            item = layout.takeAt(0)
-            widget = item.widget()
-            if widget is not None:
-                widget.deleteLater()
-            else:
-                self._clearLayout(item.layout())
-
-_plugins = [Mapper, Mapper_old]
